requirements.txt
langgraph
langchain
requests
chromadb
pydantic
python-dotenv 
SQLAlchemy 
fastapi
PyPDF2
pycryptodome